rename htmlview to htmlcore; more semantic and consistent
