+++ htmlcore/examples/html/example.html
@@ -13,9 +13,9 @@
 </head>
 
 <body>
-    <h1 style="color:blue">Gidom</h1>
+    <h1 style="color:blue">htmlcore</h1>
     <h3 id="my-heading">Made with <i>Cogent Core</i></h3>
-    <p class="red">This is a sample <b>HTML</b> <u>document</u> displayed using <code>gidom</code>.</p>
+    <p class="red">This is a sample <b>HTML</b> <u>document</u> displayed using <code>htmlcore</code>.</p>
     <button>Click me!</button>
     <p>Go to the <a href="https://cogentcore.org/core">Cogent Core website</a></p>
     <input type="text">
