+++ pages/page.go
@@ -23,7 +23,7 @@ import (
 	"cogentcore.org/core/colors"
 	"cogentcore.org/core/core"
 	"cogentcore.org/core/events"
-	"cogentcore.org/core/htmlview"
+	"cogentcore.org/core/htmlcore"
 	"cogentcore.org/core/pages/wpath"
 	"cogentcore.org/core/styles"
 	"cogentcore.org/core/system"
@@ -38,8 +38,8 @@ type Page struct {
 	// Source is the filesystem in which the content is located.
 	Source fs.FS
 
-	// Context is the page's [htmlview.Context].
-	Context *htmlview.Context `set:"-"`
+	// Context is the page's [htmlcore.Context].
+	Context *htmlcore.Context `set:"-"`
 
 	// The history of URLs that have been visited. The oldest page is first.
 	History []string `set:"-"`
@@ -66,7 +66,7 @@ var saveWebURL func(u string)
 
 func (pg *Page) Init() {
 	pg.Frame.Init()
-	pg.Context = htmlview.NewContext()
+	pg.Context = htmlcore.NewContext()
 	pg.Context.OpenURL = func(url string) {
 		pg.OpenURL(url, true)
 	}
@@ -273,7 +273,7 @@ func (pg *Page) OpenURL(rawURL string, addToHistory bool) {
 
 	fr := pg.FindPath("splits/body").(*core.Frame)
 	fr.DeleteChildren()
-	err = htmlview.ReadMD(pg.Context, fr, b)
+	err = htmlcore.ReadMD(pg.Context, fr, b)
 	if err != nil {
 		core.ErrorSnackbar(pg, err, "Error loading page")
 		return
